controladores funcionando
