--- src/controllers/__init__.py
@@ -1,5 +1,0 @@
-from flask import Blueprint
-
-weather_blueprint = Blueprint('weather_blueprint', __name__)
-
-from . import processjson
