+++ templates/result_city.html
@@ -69,123 +69,3 @@
       </div>
     </div>
 {%endblock%}
-
-<!-- 
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="../static/css/style.css">
-    <link rel="stylesheet" href="../static/css/result.css">
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-4bw+/aepP/YC94hEpVNVgiZdgIC5+VKNBQNGCHeKRQN+PtmoHDEXuppvnDJzQIu9" crossorigin="anonymous">
-    <title>{%block title%}{% endblock %}</title>
-</head>
-<body style="background-color: #dbf2f9;">
-Background image
-    <section class="image-background">
-      <div class="sun">
-        <img width="320px" src="../static/images/744-removebg-preview.png">
-      </div>
-      <div class="cloud1">
-        <img width="250px" src="../static/images/cloud.png"\>
-      </div>
-      <div class="cloud2">
-        <img width="270px" src="../static/images/cloud.png"\>
-      </div>
-      <div class="cloud3">
-        <img width="300px" src="../static/images/cloud.png"\>
-      </div>
-      <div class="cloud4">
-        <img width="250px" src="../static/images/cloud.png"\>
-      </div>
-      <div class="cloud5">
-        <img width="310px" src="../static/images/cloud.png"\>
-      </div>
-      <div class="cloud6">
-        <img width="310px" src="../static/images/cloud.png"\>
-      </div>
-      <div class="grass">
-        <img width="2000px" height="500px" src="../static/images/lovepik-lawn-png-image_401317000_wh1200-removebg-preview.png"\>
-      </div>
-      <div class="capybara">
-        <img width="300px" src="../static/images/photo_2023-09-16_16-23-28-removebg-preview.png">
-      </div>
-      <div class="capybara2">
-        <img width="300px" src="../static/images/m_capylentes-removebg-preview.png">
-      </div>
-      <div class="capybara3">
-        <img width="300px" src="../static/images/mandarina_capy-removebg-preview.png">
-      </div>
-    </section>
-    <div class="card align-items-center container d-flex text-center" style="width: 60rem;">
-      <div class="card-body">
-      <div class="container">
-        <div class="row">
-          <div class="col-15" id="card-title">{{city}}</div>
-        </div>
-        <div class="row">
-          <div class="col space">
-            <div>
-              <i>
-                <figure>
-                  <img class ="icon-weather" src="../static/images/cloud.png">
-                </figure>
-              </i>
-              <h3>{{weather_type}}</h3>
-            </div>
-          </div>
-          <div class="col">
-            <figure>
-              <img id="capybara" src="../static/images/mandarina_capy-removebg-preview.png"\>
-            </figure>
-          </div>
-          <div class="col temp_p">
-            <h3>{{temp_p}}</h3>
-          </div>
-          <div class="col">
-              <h3>Max. Tem.</h3>
-              <p>{{max_temp}}</p>
-              <h3>Min. Tem.</h3>
-              <p>{{min_temp}}</p>
-            </div>
-          </div>
-        </div>
-        <div class="weather-table">
-          <div class="row">
-            <div class="col">
-              <h4>Feels like</h4>
-              <p>{{feels_like}}</p>
-            </div>
-            <div class="col">
-              <h4>Humidity</h4>
-              <p>{{humidity}}</p>
-            </div>
-            <div class="col">
-              <h4>Rain</h4>
-              <p>{{rain}}</p>
-            </div>
-          </div>
-          <div class="row second-row ">
-            <div class="col">
-              <h4>Wind</h4>
-              <p>{{wind}}</p>
-            </div>
-            <div class="col">
-              <h4>Clouds</h4>
-              <p>{{clouds}}</p>
-            </div>
-            <div class="col">
-              <h4>Pressure</h4>
-              <p>{{pressure}}</p>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js" integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.8/dist/umd/popper.min.js" integrity="sha384-I7E8VVD/ismYTF4hNIPjVp/Zjvgyol6VFvRkX/vR+Vc4jQkC+hVqc2pM8ODewa9r" crossorigin="anonymous"></script>
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.min.js" integrity="sha384-Rx+T1VzGupg4BHQYs2gCW9It+akI2MM/mndMCy36UVfodzcJcF0GGLxZIzObiEfa" crossorigin="anonymous"></script>
-</body>
-</html> -->
